@@ -4,97 +4,87 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace MeadowBleLed.Controller
+namespace MeadowBleLed.Controller;
+
+public class LedController
 {
-    public class LedController
+    RgbPwmLed rgbPwmLed;
+
+    Task animationTask = null;
+    CancellationTokenSource cancellationTokenSource = null;
+
+    public LedController()
     {
-        private static readonly Lazy<LedController> instance =
-            new Lazy<LedController>(() => new LedController());
-        public static LedController Instance => instance.Value;
+        rgbPwmLed = new RgbPwmLed(
+            redPwmPin: MeadowApp.Device.Pins.D12,
+            greenPwmPin: MeadowApp.Device.Pins.D11,
+            bluePwmPin: MeadowApp.Device.Pins.D10);
+    }
 
-        RgbPwmLed rgbPwmLed;
+    void Stop()
+    {
+        rgbPwmLed.StopAnimation();
+        cancellationTokenSource?.Cancel();
+    }
 
-        Task animationTask = null;
-        CancellationTokenSource cancellationTokenSource = null;
+    public void SetColor(Color color)
+    {
+        Stop();
+        rgbPwmLed.SetColor(color);
+    }
 
-        private LedController()
+    public void TurnOn()
+    {
+        Stop();
+        rgbPwmLed.SetColor(GetRandomColor());
+        rgbPwmLed.IsOn = true;
+    }
+
+    public void TurnOff()
+    {
+        Stop();
+        rgbPwmLed.IsOn = false;
+    }
+
+    public void StartBlink()
+    {
+        rgbPwmLed.StartBlink(GetRandomColor());
+    }
+
+    public void StartPulse()
+    {
+        rgbPwmLed.StartPulse(GetRandomColor());
+    }
+
+    public void StartRunningColors()
+    {
+        rgbPwmLed.StopAnimation();
+
+        animationTask = new Task(async () =>
         {
-            Initialize();
-        }
+            cancellationTokenSource = new CancellationTokenSource();
+            await StartRunningColors(cancellationTokenSource.Token);
+        });
+        animationTask.Start();
+    }
 
-        private void Initialize()
+    protected async Task StartRunningColors(CancellationToken cancellationToken)
+    {
+        while (true)
         {
-            rgbPwmLed = new RgbPwmLed(
-                redPwmPin: MeadowApp.Device.Pins.D12,
-                greenPwmPin: MeadowApp.Device.Pins.D11,
-                bluePwmPin: MeadowApp.Device.Pins.D10);
-        }
-
-        void Stop()
-        {
-            rgbPwmLed.StopAnimation();
-            cancellationTokenSource?.Cancel();
-        }
-
-        public void SetColor(Color color)
-        {
-            Stop();
-            rgbPwmLed.SetColor(color);
-        }
-
-        public void TurnOn()
-        {
-            Stop();
-            rgbPwmLed.SetColor(GetRandomColor());
-            rgbPwmLed.IsOn = true;
-        }
-
-        public void TurnOff()
-        {
-            Stop();
-            rgbPwmLed.IsOn = false;
-        }
-
-        public void StartBlink()
-        {
-            rgbPwmLed.StartBlink(GetRandomColor());
-        }
-
-        public void StartPulse()
-        {
-            rgbPwmLed.StartPulse(GetRandomColor());
-        }
-
-        public void StartRunningColors()
-        {
-            rgbPwmLed.StopAnimation();
-
-            animationTask = new Task(async () =>
+            if (cancellationToken.IsCancellationRequested)
             {
-                cancellationTokenSource = new CancellationTokenSource();
-                await StartRunningColors(cancellationTokenSource.Token);
-            });
-            animationTask.Start();
-        }
-
-        protected async Task StartRunningColors(CancellationToken cancellationToken)
-        {
-            while (true)
-            {
-                if (cancellationToken.IsCancellationRequested)
-                {
-                    break;
-                }
-
-                rgbPwmLed.SetColor(GetRandomColor());
-                await Task.Delay(1000);
+                break;
             }
-        }
 
-        protected Color GetRandomColor()
-        {
-            var random = new Random();
-            return Color.FromHsba((float)random.NextDouble(), 1, 1);
+            rgbPwmLed.SetColor(GetRandomColor());
+            await Task.Delay(1000);
         }
+    }
+
+    protected Color GetRandomColor()
+    {
+        var random = new Random();
+        return Color.FromHsba((float)random.NextDouble(), 1, 1);
     }
 }
